==> submission/XXXXXXXX-FamilyName/Notebooks/02-Data preparation 1.ipynb <==
## Data preparation 1
in this notebook we show how we perform some manipulation on the full dataset in order to prepare it to feed the LSTM

the code for performing the data preparation on the test dataset is exactly the same

first load the libraries

import pandas as pd

pd.options.display.max_columns = 100
import numpy as np

then load the dataset

dataset = pd.read_csv('train_after_preprocess.csv')

at this point we encoded the events in a different way, we add a column for each possible event and we add a 1 in the column corresponding with the event or events (there may be simultaneous events) occuring in that quarter hour, 0 otherwise


# compute a list of all the possible events occuring
list_of_events = dataset[['EVENT_TYPE1', 'EVENT_TYPE2', 'EVENT_TYPE3', 'EVENT_TYPE4']].stack().unique()
# list_of_events = ['NO_EVENT', 'Veicolo_in_avaria', 'Ostacolo_in_carreggiata', 'Regimazione_delle_acque','Manutenzione_opere_in_verde', 'Meteo', 'Gestione_viabilita', 'extended_accident', 'Segnaletica_orizzontale','Allarme', 'Segnaletica_verticale', 'Pavimentazione', 'Barriere', 'Opera_arte', 'Calamita_naturale']

print(list_of_events)
print(len(list_of_events))

# delete no event since we don't want a column with no event or nan values
list_of_events.remove('NO_EVENT')


def add_columns_to_dataframe(dataframe, list_of_events):
    for event in list_of_events:
        dataframe[event] = np.int8(0)
    return dataframe

# add columns to the dataset, one for each event (for now all with zeros)
dataset = add_columns_to_dataframe(dataset, list_of_events)

def add_ones(dataframe):
    for index, row in dataframe.iterrows():
        print(index)
        for i in range(1, 5):
            if (row['EVENT_TYPE' + str(i)] != 'NO_EVENT'):
                event_type = row['EVENT_TYPE' + str(i)]
                dataframe.at[index, event_type] = np.int8(1)
    return dataframe

dataset = add_ones(dataset)

print(dataset.info())

dataset.to_csv('train_final.csv', encoding='utf-8', index=False)
