==> submission/XXXXXXXX-FamilyName/Notebooks/03- Data preparation 2.ipynb <==
## Data preparation 2
in this notebook we keep the manipulation on the full dataset in order to prepare it to feed the rnn

libraries import

import pandas as pd
pd.options.display.max_columns = 100

import numpy as np

import matplotlib.pyplot as plt
# matplotlib inline
import warnings
warnings.filterwarnings("ignore")

then we load the test and train data
here we used a dictionary to force the loading of the events columns ( with only 0 and 1) to be int8 and not int64( default)
in order to reduce the memory space

dataframe_dict = pd.read_csv('dict_to_cast_events_when_loading.csv',delimiter = ";")
dictionary = {}
for index, row in dataframe_dict.iterrows():
    dictionary[row['EVENTO']] = row['TIPO']
print(dictionary)

train = pd.read_csv('train_final.csv', dtype=dictionary)
test = pd.read_csv('test_2019_final.csv', dtype=dictionary)

print(train.info())
print(test.info())

at this point we perform the one-hot-encoding on the weather field, considering all the possible weather conditions in the
train and test set together

df_temp_train = train[['WEATHER']]
df_temp_test = test[['WEATHER']]

df_temp_train['train'] = 1
df_temp_test['train'] = 0

df_combined = pd.concat([df_temp_train, df_temp_test])

df_combined_encoded = pd.get_dummies(df_combined['WEATHER'])

df_combined = pd.concat([df_combined, df_combined_encoded], axis=1)

train_encoded = df_combined[df_combined["train"] == 1]
test_encoded = df_combined[df_combined["train"] == 0]

train_encoded.drop(['train'], axis=1, inplace=True)
train_encoded.drop(['WEATHER'], axis=1, inplace=True)
test_encoded.drop(['train'], axis=1, inplace=True)
test_encoded.drop(['WEATHER'], axis=1, inplace=True)

print(train_encoded.info())
print(test_encoded.info())

now we can delete the columns with reduntant information such the weather and the event_detail because their values have been 
encoded, the km and key because there is already key_2 attribute, and id of the station

train.drop('WEATHER', axis=1, inplace=True)
train.drop('TEMPERATURE', axis=1, inplace=True)
train.drop('MIN_TEMPERATURE', axis=1, inplace=True)
train.drop('MAX_TEMPERATURE', axis=1, inplace=True)
train.drop('ID', axis=1, inplace=True)
train.drop('EVENT_TYPE1', axis=1, inplace=True)
train.drop('EVENT_TYPE2', axis=1, inplace=True)
train.drop('EVENT_TYPE3', axis=1, inplace=True)
train.drop('EVENT_TYPE4', axis=1, inplace=True)
train.drop('KEY', axis=1, inplace=True)
train.drop('KM', axis=1, inplace=True)

test.drop('WEATHER', axis=1, inplace=True)
test.drop('TEMPERATURE', axis=1, inplace=True)
test.drop('MIN_TEMPERATURE', axis=1, inplace=True)
test.drop('MAX_TEMPERATURE', axis=1, inplace=True)
test.drop('ID', axis=1, inplace=True)
test.drop('EVENT_TYPE1', axis=1, inplace=True)
test.drop('EVENT_TYPE2', axis=1, inplace=True)
test.drop('KEY', axis=1, inplace=True)
test.drop('KM', axis=1, inplace=True)

train = pd.concat([train, train_encoded], axis=1)
test = pd.concat([test, test_encoded], axis=1)

print(train.info())
print(train.head(10))

print(test.info())
print(test.head(10))

save the data

train.to_csv('train_final_after_weather_encoding.csv')
test.to_csv('test_final_after_weather_encoding.csv')
